requirements.txt
tqdm
selenium
pandas
python-docx
spire.doc
scrapy
opencv-python
requests
numpy
pudb
unidecode